Remove the function-arity cap: unbounded positional parameters
Functions no longer have a hard parameter-count ceiling. A defn, fn, or
extern-c may declare an arbitrary number of positional parameters, bounded only
by uint32_t -- matching the emitted C, which has no limit of its own. The
previous change had raised the mechanical cap from 16 to 64; this removes it.

Core restructure (arbitrary-fn-arity-plan Phase 2):

  - Out-of-line per-arg storage. Type.as.fn.arg_kinds / arg_flags are arena
    pointers (from a process-lifetime global type arena, tur_fn_args_alloc)
    instead of inline [MAX_FN_ARITY] arrays; fn arity is uint32_t. sizeof(Type)
    drops from 304 to 128 -- the fn variant no longer dominates the union, so
    deep by-value-Type codegen recursion gains headroom. The few sites that
    rebuild a fn type at a different arity (dict/env injection, poly and
    struct-field instantiation, forward-decl growth) allocate fresh arrays
    rather than mutate a shared one.

  - FnDef.n_params / ExternC.n_params / EmitCtx.n_fn_params and the arity/param
    loop counters and count locals across elab, emit, and the passes were
    widened from uint8_t to uint32_t (a uint8_t FnDef.n_params was silently
    truncating 300 -> 44). The >= MAX_FN_ARITY hard-reject guards are gone.

  - Per-param buffers are arena/count-sized, not fixed [MAX_FN_ARITY]:
    elaboration param scratch (defn/fn/extern-c), fwd_decl_scan_params, the
    emit param/thunk/needs-box buffers, ABI-instantiation scratch, and pbp
    (pass-by-ptr) params, which is now a realloc-grown array on EmitCtx.

  - poly_arg_mask / poly_agg_arg_mask are uint64_t indexed via ARG_IDX_BIT(i)
    (0 for i >= 64) so ordinary high-arity calls never hit an undefined shift.

  - Fixed a latent bug the smaller Type union exposed: effect_check read
    .as.fn.effect_row off a higher-order param that may be TY_FORALL, not TY_FN,
    yielding a garbage pointer; now guarded on kind == TY_FN.

Exceeding the historical 16 params remains a TUR-W0041 lint nudge, not an error.
MAX_FN_ARITY (64) survives only as the default size of a few internal codegen
fast-path buffers (ABI specialization, interpreter effect-perform args), each of
which falls back gracefully / errors cleanly for wider functions.

Verified end to end (compiled) at 100/300/500/1000 ordinary params, an 80-param
generic function, and a 70-param struct-by-value function. New fixture
fn-arity-100-param. Full suite green (2116 passed, 0 failed).

+++ src/compiler/elab_effects.c
@@ -561,7 +561,7 @@ Expr *elab_defeffect(Elab *e, const Form *call) {
      * loses the def.  Entries stay NULL for scalars; the whole array is NULL if no
      * param is an aggregate (set after the loop). */
     const Type **param_full = arena_alloc(e->arena, (n_params ? n_params : 1) * sizeof(const Type *));
-    for (uint8_t j = 0; j < n_params; j++) param_full[j] = NULL;
+    for (uint32_t j = 0; j < n_params; j++) param_full[j] = NULL;
     bool any_agg_param = false;
 
     {
@@ -828,7 +828,7 @@ Expr *elab_perform(Elab *e, const Form *call) {
     /* Parse arguments */
     uint8_t n_args = effect_call_f->as.list.len - 1;
     Expr **args = arena_alloc(e->arena, n_args * sizeof(Expr *));
-    for (uint8_t i = 0; i < n_args; i++) {
+    for (uint32_t i = 0; i < n_args; i++) {
         args[i] = elab_form(e, effect_call_f->as.list.items[i + 1]);
         if (!args[i]) return NULL;
     }
@@ -945,7 +945,7 @@ Expr *elab_handle(Elab *e, const Form *call) {
         cases[i].n_params = params_f->as.list.len;
         cases[i].param_names = arena_alloc(e->arena, cases[i].n_params * sizeof(const Symbol *));
         cases[i].param_bindings = arena_alloc(e->arena, cases[i].n_params * sizeof(Binding *));
-        for (uint8_t j = 0; j < cases[i].n_params; j++) {
+        for (uint32_t j = 0; j < cases[i].n_params; j++) {
             Form *param_f = params_f->as.list.items[j];
             if (param_f->tag != F_SYM) {
                 diag_emit(DIAG_ERROR, param_f->span,
@@ -1011,7 +1011,7 @@ Expr *elab_handle(Elab *e, const Form *call) {
         e->scope = &handler_scope;
         
         /* Create bindings for each parameter */
-        for (uint8_t j = 0; j < cases[i].n_params; j++) {
+        for (uint32_t j = 0; j < cases[i].n_params; j++) {
             /* Use the effect's declared param type if available, else TY_INT.
              * Tier C: prefer the full param Type (carries the real aggregate def)
              * so field access on the param binding elaborates. */
@@ -1071,7 +1071,7 @@ Expr *elab_handle(Elab *e, const Form *call) {
             /* Build the "local params" list: effect params + k (not free vars) */
             uint8_t n_hparams = (uint8_t)(cases[i].n_params + 1);
             Binding **hparams = arena_alloc(e->arena, n_hparams * sizeof(Binding *));
-            for (uint8_t j = 0; j < cases[i].n_params; j++)
+            for (uint32_t j = 0; j < cases[i].n_params; j++)
                 hparams[j] = cases[i].param_bindings[j];
             hparams[cases[i].n_params] = kb;
             uint32_t n_caps = 0;
@@ -1214,7 +1214,7 @@ Expr *elab_handler_lit(Elab *e, const Form *call) {
     cases[0].n_params = params_f->as.list.len;
     cases[0].param_names = arena_alloc(e->arena, cases[0].n_params * sizeof(const Symbol *));
     cases[0].param_bindings = arena_alloc(e->arena, cases[0].n_params * sizeof(Binding *));
-    for (uint8_t j = 0; j < cases[0].n_params; j++) {
+    for (uint32_t j = 0; j < cases[0].n_params; j++) {
         Form *param_f = params_f->as.list.items[j];
         if (param_f->tag != F_SYM) {
             diag_emit(DIAG_ERROR, param_f->span, "handler literal: parameter name must be a symbol");
@@ -1265,7 +1265,7 @@ Expr *elab_handler_lit(Elab *e, const Form *call) {
     Scope *saved_scope = e->scope;
     e->scope = &handler_scope;
 
-    for (uint8_t j = 0; j < cases[0].n_params; j++) {
+    for (uint32_t j = 0; j < cases[0].n_params; j++) {
         TypeKind pk = (eff && j < eff->constructor->n_params)
             ? eff->constructor->param_types[j] : TY_INT;
         Type ptype = type_from_kind(pk);
@@ -1300,7 +1300,7 @@ Expr *elab_handler_lit(Elab *e, const Form *call) {
     if (cases[0].cont_kind == CK_MULTISHOT && cases[0].body) {
         uint8_t n_hparams = (uint8_t)(cases[0].n_params + 1);
         Binding **hparams = arena_alloc(e->arena, n_hparams * sizeof(Binding *));
-        for (uint8_t j = 0; j < cases[0].n_params; j++)
+        for (uint32_t j = 0; j < cases[0].n_params; j++)
             hparams[j] = cases[0].param_bindings[j];
         hparams[cases[0].n_params] = kb;
         uint32_t n_caps = 0;
